g2d: fix checkpatch warning and errors
Suppressed all warning, check and error messages to all source files of
G2D driver using the following command:
"scripts/checkpatch.pl --strict --show-types --ignore GERRIT_CHANGE_ID
--ignore FILE_PATH_CHANGES"

Some checks and warnings are ignored like BUG() for some possible future
design flaws and documentation for device tree compatible strings.

(cherry picked from commit 914efef4958c22101b186be6761899a844ebed50)
Bug: 165562344

+++ drivers/gpu/exynos/g2d/g2d_command.h
@@ -33,9 +33,9 @@ bool g2d_validate_target_commands(struct g2d_device *g2d_dev,
 				  struct g2d_task *task);
 size_t g2d_get_payload(struct g2d_reg cmd[], const struct g2d_fmt *fmt,
 		       u32 flags, unsigned long caps);
-size_t g2d_get_payload_index(struct g2d_reg cmd[], const struct g2d_fmt *fmt,
-			     unsigned int idx, unsigned int buffer_count,
-			     unsigned long caps, u32 flags);
+unsigned int g2d_get_payload_index(struct g2d_reg cmd[], const struct g2d_fmt *fmt,
+				   unsigned int idx, unsigned int buffer_count,
+				   unsigned long caps, u32 flags);
 bool g2d_prepare_source(struct g2d_task *task,
 			struct g2d_layer *layer, int index);
 bool g2d_prepare_target(struct g2d_task *task);
